UserServices and connection with database
The user registration function is responsible for creating a new user account and storing the user's information in the database. This function should include validation checks to ensure the provided user details are accurate and meet certain criteria as ( firstname , lastname ,email ,password).......
and The user authentication function verifies the user's identity during the login process. It checks the provided credentials against the stored information in the database .